+++ playlet-web/src/lib/VideoFeed/ChannelCell.svelte
@@ -8,6 +8,9 @@
   export let authorThumbnails: any[] | undefined = undefined;
   export let channelHandle: string | undefined = undefined;
   export let subCount: number | undefined = undefined;
+  export let subCountText: string | undefined = undefined;
+  export let videoCountText: string | undefined = undefined;
+  export let authorSubTitle: string | undefined = undefined;
 
   // svelte-ignore unused-export-let
   export let type: string = undefined;
@@ -15,7 +18,17 @@
   let modal;
 
   function getSubCountText() {
+    if (authorSubTitle) {
+      return authorSubTitle;
+    }
+
     if (isNaN(subCount) || subCount < 0) {
+      if (videoCountText) {
+        return videoCountText;
+      }
+      if (subCountText) {
+        return subCountText;
+      }
       return "";
     }
 
@@ -32,7 +45,7 @@
   <div class="card card-compact bg-base-100 shadow-xl border border-neutral">
     <ChannelThumbnail bind:author bind:authorThumbnails />
     <div class="card-body">
-      <h3 class="card-title text-base line-clamp-1 min-h-8">{author}</h3>
+      <h3 class="card-title text-base line-clamp-2 min-h-8">{author}</h3>
       <div>{getSubCountText()}</div>
       <div>{channelHandle || ""}</div>
     </div>