+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Support for auto-generated channels in Playlet backend
 
+### Fixed
+
+- Channels sub count in web app
+
 ### Removed
 
 - [YouTube.js](https://github.com/LuanRT/YouTube.js) from web app. It is no longer needed, so this reduces app size.